First version of recording last status update. Also fix date serialization
