+++ hrm-domain/hrm-service/service-tests/case/caseValidation.ts
@@ -14,9 +14,9 @@
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
 
-import { WELL_KNOWN_CASE_SECTION_NAMES } from '../src/case/caseService';
-import { NewContactRecord } from '../src/contact/sql/contact-insert-sql';
-import { ContactRawJson, CreateContactPayload } from '../src/contact/contactService';
+import { CaseService, WELL_KNOWN_CASE_SECTION_NAMES } from '../../src/case/caseService';
+import { NewContactRecord } from '../../src/contact/sql/contact-insert-sql';
+import { ContactRawJson, CreateContactPayload } from '../../src/contact/contactService';
 
 declare global {
   namespace jest {
@@ -67,10 +67,10 @@ export const without = (original, ...property) => {
 };
 
 export const convertCaseInfoToExpectedInfo = (
-  input: any,
+  input: Partial<CaseService>,
   accountSid: string | null = null,
-) => {
-  if (!input || !input.info) return { ...input };
+): CaseService => {
+  if (!input || !input.info) return { ...input } as CaseService;
   const expectedCase = {
     ...input,
     info: { ...input.info },
@@ -93,7 +93,7 @@ export const convertCaseInfoToExpectedInfo = (
       }
     });
   }
-  return expectedCase;
+  return expectedCase as CaseService;
 };
 
 export const validateCaseListResponse = (actual, expectedCaseAndContactModels, count) => {
