+++ hrm-domain/hrm-service/service-tests/db-tests/audit_trigger.test.ts
@@ -21,7 +21,7 @@
 
 import { db } from '../../src/connection-pool';
 
-import '../case-validation';
+import '../case/caseValidation';
 
 const workerSid = 'WK-worker-sid';
 const anotherWorkerSid = 'WK-another-worker-sid';
